+++ lib/Elastica/QueryBuilder/DSL/Collapse.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Elastica\QueryBuilder\DSL;
+
+use Elastica\Query\InnerHits;
+use Elastica\QueryBuilder\DSL;
+
+/**
+ * Class Collapse.
+ *
+ *
+ * @see https://www.elastic.co/guide/en/elasticsearch/reference/6.7/search-request-collapse.html
+ */
+class Collapse implements DSL
+{
+    /**
+     * @return string
+     */
+    public function getType(): string
+    {
+        return self::TYPE_COLLAPSE;
+    }
+
+    /**
+     * @return InnerHits
+     */
+    public function inner_hits(): InnerHits
+    {
+        return new InnerHits();
+    }
+}